Add rudimentary C++ code generation
This code generates enums (as namespaces with constexpr values),
and messages. Messages get getters and setters generated for all
members, and a encoder and a decoder function.

The encoders/decoders only handle fixed width types at the moment though.
Support for tagged members and dynamic width arrays (type[] and string members)
is still not present at the moment.