meta: only check for the C++ compiler when needed
This avoids the check when it is unnecessary, as that might be unwanted
in some cases. For instance, using `subproject("bragi")` would trigger
the check, even if the intent is only to get the dependency to pull in
the support headers, where no compiler is necessary. This becomes
relevant when building `mlibc-headers`, as that is required for building
a cross-gcc, and as such has no cross-gcc available.